Add frame latency tracker
Haha! Numbers!

+++ Source/Bristlecone/Public/FBristleconePacket.h
@@ -35,7 +35,7 @@ public:
 		return &packet;
 	}
 
-	long long GetSendTimeStamp() const {
+	long GetSendTimeStamp() const {
 		return packet.GetTransferTime();
 	}
